ref(symcache): Move all dwarf logic to own module
